refactor: Split index.dart into individual files
Splits the `mapper/lib/src/model/index.dart` file into individual files. This improves maintainability and readability of the code.

- Created a new directory `mapper/lib/src/model/index`.
- Moved `DeserializationOptions`, `SerializationOptions`, `DeserializationContext`, `SerializationContext`, `PropertyDescriptor`, `ProcessedObjectDescriptor`, and other definitions into their own files within the new directory.
- Replaced the original `index.dart` with a barrel file that exports all the new files, ensuring no breaking changes for existing importers.
