+++ mapper/lib/src/model/converters/regexp_converter.dart
@@ -1,0 +1,17 @@
+import '../index.dart';
+import 'base_converter.dart';
+
+const regExpConverter = RegExpConverter();
+
+/// RegExp converter
+class RegExpConverter implements ICustomConverter<RegExp?> {
+  const RegExpConverter() : super();
+
+  @override
+  RegExp? fromJSON(dynamic jsonValue, DeserializationContext context) =>
+      jsonValue is String ? RegExp(jsonValue) : jsonValue;
+
+  @override
+  dynamic toJSON(RegExp? object, SerializationContext context) =>
+      object?.pattern;
+}
