+++ mapper/lib/src/model/index/conversion_direction.dart
@@ -1,0 +1,1 @@
+enum ConversionDirection { fromJson, toJson }
