+++ mapper/lib/src/model/index/injectable_values.dart
@@ -1,0 +1,1 @@
+typedef InjectableValues = Map<String, dynamic>;
